third_party/blink/web_tests/fast/css/invalidation/font-face-load.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
    .x { color: green }
</style>
<div id="x">
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
    description("Check that FontFaceSet.load() does not trigger a style recalc");

    x.offsetTop;
    x.className = "x";
    document.fonts.load("40px dummy");

    if (window.internals)
        shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

    shouldBeEqualToString("getComputedStyle(x).color", "rgb(0, 128, 0)");
</script>
