third_party/blink/web_tests/fast/css/invalidation/font-face-load-expected.txt

Check that FontFaceSet.load() does not trigger a style recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS getComputedStyle(x).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
